+++ Plugin.cs
@@ -7,7 +7,12 @@ using GeneralImprovements.Utilities;
 using HarmonyLib;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
+using System.IO;
 using System.Linq;
+using System.Reflection;
+using System.Text.RegularExpressions;
+using UnityEngine;
 using UnityEngine.InputSystem;
 
 namespace GeneralImprovements
@@ -20,6 +25,10 @@ namespace GeneralImprovements
         private const string ExtraMonitorsSection = "ExtraMonitors";
         public static ConfigEntry<bool> UseBetterMonitors { get; private set; }
         public static ConfigEntry<bool> ShowBlueMonitorBackground { get; private set; }
+        public static ConfigEntry<string> MonitorBackgroundColor { get; private set; }
+        public static Color MonitorBackgroundColorVal { get; private set; }
+        public static ConfigEntry<string> MonitorTextColor { get; private set; }
+        public static Color MonitorTextColorVal { get; private set; }
         public static ConfigEntry<bool> ShowBackgroundOnAllScreens { get; private set; }
         public static ConfigEntry<int> ShipProfitQuotaMonitorNum { get; private set; }
         public static ConfigEntry<int> ShipDeadlineMonitorNum { get; private set; }
@@ -34,6 +43,8 @@ namespace GeneralImprovements
         public static ConfigEntry<bool> CenterAlignMonitorText { get; private set; }
         public static ConfigEntry<int> ShipInternalCamSizeMultiplier { get; private set; }
         public static ConfigEntry<int> ShipInternalCamFPS { get; private set; }
+        public static ConfigEntry<int> ShipExternalCamSizeMultiplier { get; private set; }
+        public static ConfigEntry<int> ShipExternalCamFPS { get; private set; }
 
         private const string FixesSection = "Fixes";
         public static ConfigEntry<bool> FixInternalFireExits { get; private set; }
@@ -65,6 +76,7 @@ namespace GeneralImprovements
         public static ConfigEntry<string> FreeRotateKey { get; private set; }
         public static ConfigEntry<string> CounterClockwiseKey { get; private set; }
         public static ConfigEntry<bool> ShipMapCamDueNorth { get; private set; }
+        public static ConfigEntry<bool> SpeakerPlaysIntroVoice { get; private set; }
 
         private const string TeleportersSection = "Teleporters";
         public static ConfigEntry<int> RegularTeleporterCooldown { get; private set; }
@@ -73,6 +85,7 @@ namespace GeneralImprovements
         private const string TerminalSection = "Terminal";
         public static ConfigEntry<int> TerminalHistoryItemCount { get; private set; }
         public static ConfigEntry<bool> TerminalFastCamSwitch { get; private set; }
+        public static ConfigEntry<bool> LockCameraAtTerminal { get; private set; }
 
         private const string ToolsSection = "Tools";
         public static ConfigEntry<string> ScannableTools { get; private set; }
@@ -99,8 +112,10 @@ namespace GeneralImprovements
             // Extra monitors
             string numDesc = "0 = DISABLED. If UseBetterMonitors = True, 1-6 are the top, 7-12 are the bottom, and 13-14 are the big ones beside the terminal. Otherwise, 1-4 are the top, and 5-8 are on the bottom.";
             UseBetterMonitors = Config.Bind(ExtraMonitorsSection, nameof(UseBetterMonitors), false, "If set to true, uses 12 fully customizable and integrated monitors instead of the 8 vanilla ones with overlays.");
-            ShowBlueMonitorBackground = Config.Bind(ExtraMonitorsSection, nameof(ShowBlueMonitorBackground), true, "If set to true, keeps the vanilla blue backgrounds on the extra monitors. Set to false for black.");
-            ShowBackgroundOnAllScreens = Config.Bind(ExtraMonitorsSection, nameof(ShowBackgroundOnAllScreens), false, "If paired with ShowBlueMonitorBackground, Shows the background color on ALL monitors when they are on, not just used ones.");
+            ShowBlueMonitorBackground = Config.Bind(ExtraMonitorsSection, nameof(ShowBlueMonitorBackground), true, "If set to true and NOT using UseBetterMonitors, keeps the vanilla blue backgrounds on the extra monitors. Set to false to hide.");
+            MonitorBackgroundColor = Config.Bind(ExtraMonitorsSection, nameof(MonitorBackgroundColor), "160959", "The hex color code of what the backgrounds of the monitors should be. A recommended value close to black is 050505.");
+            MonitorTextColor = Config.Bind(ExtraMonitorsSection, nameof(MonitorTextColor), "00FF2C", "The hex color code of what the text on the monitors should be.");
+            ShowBackgroundOnAllScreens = Config.Bind(ExtraMonitorsSection, nameof(ShowBackgroundOnAllScreens), false, "If set to true, will show the MonitorBackgroundColor on ALL monitors when they are on, not just used ones.");
             ShipProfitQuotaMonitorNum = Config.Bind(ExtraMonitorsSection, nameof(ShipProfitQuotaMonitorNum), 5, new ConfigDescription($"Which monitor to show the profit quota on. Vanilla = 5. {numDesc}", new AcceptableValueRange<int>(0, 14)));
             ShipDeadlineMonitorNum = Config.Bind(ExtraMonitorsSection, nameof(ShipDeadlineMonitorNum), 6, new ConfigDescription($"Which monitor to show the deadline on. Vanilla = 6. {numDesc}", new AcceptableValueRange<int>(0, 14)));
             ShipTotalMonitorNum = Config.Bind(ExtraMonitorsSection, nameof(ShipTotalMonitorNum), 0, new ConfigDescription($"Displays the sum of all scrap value on the ship on the specified monitor. {numDesc}", new AcceptableValueRange<int>(0, 14)));
@@ -114,6 +129,8 @@ namespace GeneralImprovements
             CenterAlignMonitorText = Config.Bind(ExtraMonitorsSection, nameof(CenterAlignMonitorText), true, "If set to true, all small monitors in the ship will have their text center aligned, instead of left.");
             ShipInternalCamSizeMultiplier = Config.Bind(ExtraMonitorsSection, nameof(ShipInternalCamSizeMultiplier), 1, new ConfigDescription($"How many times to double the internal ship cam's resolution.", new AcceptableValueRange<int>(1, 5)));
             ShipInternalCamFPS = Config.Bind(ExtraMonitorsSection, nameof(ShipInternalCamFPS), 5, new ConfigDescription($"Limits the FPS of the internal ship cam for performance. 0 = Unrestricted.", new AcceptableValueRange<int>(0, 60)));
+            ShipExternalCamSizeMultiplier = Config.Bind(ExtraMonitorsSection, nameof(ShipExternalCamSizeMultiplier), 1, new ConfigDescription($"How many times to double the external ship cam's resolution.", new AcceptableValueRange<int>(1, 5)));
+            ShipExternalCamFPS = Config.Bind(ExtraMonitorsSection, nameof(ShipExternalCamFPS), 5, new ConfigDescription($"Limits the FPS of the external ship cam for performance. 0 = Unrestricted.", new AcceptableValueRange<int>(0, 60)));
 
             // Fixes
             FixInternalFireExits = Config.Bind(FixesSection, nameof(FixInternalFireExits), true, "If set to true, the player will face the interior of the facility when entering through a fire entrance.");
@@ -143,6 +160,7 @@ namespace GeneralImprovements
             FreeRotateKey = Config.Bind(ShipSection, nameof(FreeRotateKey), Key.LeftAlt.ToString(), new ConfigDescription("If SnapObjectsByDegrees > 0, configures which modifer key activates free rotation.", new AcceptableValueList<string>(validKeys)));
             CounterClockwiseKey = Config.Bind(ShipSection, nameof(CounterClockwiseKey), Key.LeftShift.ToString(), new ConfigDescription("If SnapObjectsByDegrees > 0, configures which modifier key spins it CCW.", new AcceptableValueList<string>(validKeys)));
             ShipMapCamDueNorth = Config.Bind(ShipSection, nameof(ShipMapCamDueNorth), false, "If set to true, the ship's map camera will rotate so that it faces north evenly, instead of showing everything at an angle.");
+            SpeakerPlaysIntroVoice = Config.Bind(ShipSection, nameof(SpeakerPlaysIntroVoice), true, "If set to true, the ship's speaker will play the introductory welcome audio on the first day.");
 
             // Teleporters
             RegularTeleporterCooldown = Config.Bind(TeleportersSection, nameof(RegularTeleporterCooldown), 10, new ConfigDescription("How many seconds to wait in between button presses for the REGULAR teleporter. Vanilla = 10.", new AcceptableValueRange<int>(0, 300)));
@@ -151,6 +169,7 @@ namespace GeneralImprovements
             // Terminal
             TerminalHistoryItemCount = Config.Bind(TerminalSection, nameof(TerminalHistoryItemCount), 20, new ConfigDescription("How many items to keep in your terminal's command history. Previous terminal commands may be navigated by using the up/down arrow keys.", new AcceptableValueRange<int>(0, 100)));
             TerminalFastCamSwitch = Config.Bind(TerminalSection, nameof(TerminalFastCamSwitch), true, "If set to true, will allow use of the left/right arrow keys to quickly cycle through radar cameras while using the terminal.");
+            LockCameraAtTerminal = Config.Bind(TerminalSection, nameof(LockCameraAtTerminal), true, "If set to true, the camera will no longer move around when moving your mouse/controller while at the terminal.");
 
             // Tools
             ScannableTools = Config.Bind(ToolsSection, nameof(ScannableTools), string.Empty, $"A comma separated list of which tools, if any, should be scannable. Accepted values: {validToolStrings}");
@@ -161,10 +180,20 @@ namespace GeneralImprovements
             ShowUIReticle = Config.Bind(UISection, nameof(ShowUIReticle), false, "If set to true, the HUD will display a small dot so you can see exactly where you are pointing at all times.");
 
             // Sanitize where needed
+            string backgroundHex = Regex.Match(MonitorBackgroundColor.Value, "([a-fA-F0-9]{6})").Groups[1].Value.ToUpper();
+            string textHex = Regex.Match(MonitorTextColor.Value, "([a-fA-F0-9]{6})").Groups[1].Value.ToUpper();
+            if (backgroundHex.Length != 6) MLS.LogWarning("Invalid hex code used for monitor background color! Reverting to default.");
+            if (textHex.Length != 6) MLS.LogWarning("Invalid hex code used for monitor text color! Reverting to default.");
+            MonitorBackgroundColor.Value = backgroundHex.Length == 6 ? backgroundHex : MonitorBackgroundColor.DefaultValue.ToString();
+            MonitorBackgroundColorVal = HexToColor(MonitorBackgroundColor.Value);
+            MonitorTextColor.Value = textHex.Length == 6 ? textHex : MonitorTextColor.DefaultValue.ToString();
+            MonitorTextColorVal = HexToColor(MonitorTextColor.Value);
+
             if (MinimumStartingMoney.Value < StartingMoneyPerPlayer.Value)
             {
                 MinimumStartingMoney.Value = StartingMoneyPerPlayer.Value;
             }
+
             var validGrabbables = new List<string>();
             string[] specifiedScannables = ScannableTools.Value.Replace(" ", "").Split(',');
             if (specifiedScannables.Any(s => s.ToUpper() == "ALL"))
@@ -191,6 +220,7 @@ namespace GeneralImprovements
                 ScannableTools.Value = string.Join(',', validGrabbables.ToArray());
             }
 
+            MigrateOldConfigValues();
             MLS.LogDebug("Configuration Initialized.");
 
             Harmony.CreateAndPatchAll(typeof(DepositItemsDeskPatch));
@@ -245,6 +275,110 @@ namespace GeneralImprovements
             GameNetworkManagerPatch.PatchNetcode();
 
             MLS.LogInfo($"{Metadata.PLUGIN_NAME} v{Metadata.VERSION} fully loaded.");
+        }
+
+        private static Color HexToColor(string hex)
+        {
+            float r = int.Parse(hex.Substring(0, 2), NumberStyles.HexNumber) / 255f;
+            float g = int.Parse(hex.Substring(2, 2), NumberStyles.HexNumber) / 255f;
+            float b = int.Parse(hex.Substring(4, 2), NumberStyles.HexNumber) / 255f;
+
+            return new Color(r, g, b);
+        }
+
+        private void MigrateOldConfigValues()
+        {
+            try
+            {
+                string[] configLines = File.ReadAllLines(Config.ConfigFilePath);
+                var ourEntries = new Dictionary<string, List<InternalConfigDef>> { { string.Empty, new List<InternalConfigDef>() } };
+                string curSection = string.Empty;
+                string curDescription = string.Empty;
+
+                // Manually read sections and entries since the config classes don't provide a way to see unused values
+                foreach (string line in configLines.Select(l => l.Trim()))
+                {
+                    if (line.StartsWith('#'))
+                    {
+                        if (line.StartsWith("##")) curDescription = line.Substring(2);
+                        continue;
+                    }
+
+                    if (line.StartsWith('[') && line.EndsWith(']'))
+                    {
+                        curSection = line.Substring(1, line.Length - 2);
+                        ourEntries.TryAdd(curSection, new List<InternalConfigDef>());
+                        continue;
+                    }
+
+                    string[] entry = line.Split('=');
+                    if (entry.Length == 2)
+                    {
+                        ourEntries[curSection].Add(new InternalConfigDef(curSection, curDescription, entry[0].Trim(), entry[1].Trim()));
+                    }
+                }
+
+                // Find definitions that are not part of our current keys and remove migrate if possible
+                bool foundOrphans = false;
+                foreach (string section in ourEntries.Keys)
+                {
+                    foreach (var entry in ourEntries[section])
+                    {
+                        if (!Config.Any(k => k.Key.Section == section && entry.Name == k.Value.Definition.Key))
+                        {
+                            MigrateSpecificValue(entry);
+                            foundOrphans = true;
+                        }
+                    }
+                }
+
+                // Manually clear the private orphans
+                if (foundOrphans)
+                {
+                    var orphans = (Dictionary<ConfigDefinition, string>)Config.GetType().GetProperty("OrphanedEntries", BindingFlags.Instance | BindingFlags.NonPublic)?.GetValue(Config);
+                    if (orphans != null)
+                    {
+                        orphans.Clear();
+                        Config.Save();
+                    }
+                    else
+                    {
+                        MLS.LogWarning("Could not clear orphaned config values when migrating old config values.");
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                MLS.LogError($"Error encountered while migrating old config values! This will not affect gameplay, but please verify your config file to ensure the settings are as you expect.\n\n{ex}");
+            }
+        }
+
+        private void MigrateSpecificValue(InternalConfigDef entry)
+        {
+            MLS.LogMessage($"Found unused config value: {entry.Name}. Migrating and removing if possible...");
+
+            switch (entry.Name)
+            {
+                default:
+                    MLS.LogDebug("No matching migration");
+                    break;
+            }
+        }
+
+        private class InternalConfigDef
+        {
+            public readonly string Section;
+            public readonly string Description;
+            public readonly string Name;
+            public readonly string Value;
+
+            public InternalConfigDef(string section, string description, string name, string value)
+            {
+                Section = section;
+                Description = description;
+                Name = name;
+                Value = value;
+            }
         }
     }
 }